Related keyphrase status neutral/orange instead of happy/green when all the results are good #4475 Remove related keyword score aggregator and move filter into base SEOScoreAggregator
