+++ packages/yoastseo/spec/scoring/scoreAggregators/SEOScoreAggregatorSpec.js
@@ -111,5 +111,17 @@ describe( "SEOScoreAggregator", () => {
 			const score = aggregator.aggregate( results );
 			expect( score ).toBe( 63 );
 		} );
+
+		it( "exclude assessments without score from aggregator", () => {
+			const results = [
+				new AssessmentResult( { score: 5 } ),
+				new AssessmentResult(),
+				new AssessmentResult( { score: 4 } ),
+				new AssessmentResult( { score: 8 } ),
+				new AssessmentResult(),
+			];
+			const score = aggregator.aggregate( results );
+			expect( score ).toBe( 63 );
+		} );
 	} );
 } );
