+++ packages/yoastseo/src/scoring/scoreAggregators/SEOScoreAggregator.js
@@ -33,6 +33,18 @@ const ScoreFactor = 9;
  */
 class SEOScoreAggregator extends ScoreAggregator {
 	/**
+	 * Returns the list of valid results.
+	 * Valid results are all results that have a score.
+	 *
+	 * @param {AssessmentResult[]} results The results to filter the valid results from.
+	 *
+	 * @returns {AssessmentResult[]} The list of valid results.
+	 */
+	getValidResults( results ) {
+		return results.filter( result => result.hasScore() );
+	}
+
+	/**
 	 * Aggregates the given assessment results into a single score.
 	 *
 	 * @param {AssessmentResult[]} results The assessment results.
@@ -40,14 +52,16 @@ class SEOScoreAggregator extends ScoreAggregator {
 	 * @returns {number} The aggregated score.
 	 */
 	aggregate( results ) {
-		const score = results.reduce( ( sum, result ) => sum + result.getScore(), 0 );
+		const validResults = this.getValidResults( results );
+
+		const score = validResults.reduce( ( sum, result ) => sum + result.getScore(), 0 );
 
 		/*
 		 * Whenever the divide by part is 0 this can result in a `NaN` value. Then 0 should be returned as fallback.
-		 * This seemed better than the if check `results.length === 0`,
+		 * This seemed better than the if check `validResults.length === 0`,
 		 * because it also protects against ScoreFactor being 0.
 		 */
-		return Math.round( ( score * ScoreScale ) / ( results.length * ScoreFactor ) ) || 0;
+		return Math.round( ( score * ScoreScale ) / ( validResults.length * ScoreFactor ) ) || 0;
 	}
 }
